break apart instance launch actions and modals
