+++ troposphere/static/js/actions/InstanceActions.js
@@ -9,7 +9,7 @@ define(function (require) {
     reboot: require('./instance/reboot').reboot,
     poll: require('./instance/poll').poll,
     launch: require('./instance/launch').launch,
-    launchIntoProject: require('./instance/launch').launchIntoProject,
+    createProjectAndLaunchInstance: require('./instance/launch').createProjectAndLaunchInstance,
     destroy: require('./instance/destroy').destroy,
     update: require('./instance/update').update,
     report: require('./instance/report').report,
