+++ troposphere/static/js/modals/InstanceModals.js
@@ -2,7 +2,8 @@ define(function (require) {
 
   return {
     createAndAddToProject: require('./instance/createAndAddToProject').createAndAddToProject,
-    destroy: require('./instance/destroy').destroy
+    destroy: require('./instance/destroy').destroy,
+    launch: require('./instance/launch').launch
   };
 
 });
